backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-rest-knox
dynaconf
hvac
loguru
requests
shortuuid
sqlparse
tzdata
validators